makeHashTable is factored out
